Authenticate Twitter API access and get bearer token

oauth_twitter_token authenticates the provided credentials against the Twitter oauth2/token endpoint and returns the authentication (bearer) token required for calls to other Twitter API endpoints (consult the Twitter API documentation on application-only authentication for details).

oauth_twitter_token(consumerKey, consumerSecret)

Arguments

consumerKey
a character string specifying the consumer key for the registered authenticating Twitter application in the Academic Research product track

consumerSecret
a character string specifying the consumer secret for the registered authenticating Twitter application in the Academic Research product track

Value

a character string representing a valid bearer token

Details

This authentication method requires a consumer key and consumer secret for a registered Twitter application (see here for details).